apply MVC pattern
