+++ App/server/controllers/product_controller.py
@@ -1,135 +1,23 @@
-from server import app
-import bcrypt
 from collections import defaultdict
-from flask import Flask, render_template, flash, redirect, url_for, request, jsonify
-from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
-import pymysql
+from flask import url_for, request, jsonify, render_template
 import os
 import random
-from server.user import get_user, create_user
-from server.products import get_products, get_products_variants, create_product
-from server.tracking import get_user_behavior_by_date
-from itertools import groupby
+from server import app
+from server.models.product_model import get_products, get_products_variants, create_product
 from werkzeug.urls import url_parse
 from werkzeug.utils import secure_filename
 
 PAGE_SIZE = 6
 ALLOWED_EXTENSIONS = set(['pdf', 'png', 'jpg', 'jpeg', 'gif'])
 
-def dir_last_updated(folder):
-    return str(max(os.path.getmtime(os.path.join(root_path, f))
-                   for root_path, dirs, files in os.walk(folder)
-                   for f in files))
+@app.route('/admin/product.html', methods=['GET'])
+def admin_product():
+    return render_template('product_create.html')
 
-def get_hashed_password(plain_text_password):
-    return bcrypt.hashpw(plain_text_password.encode('utf8'), bcrypt.gensalt())
-
-def check_password(plain_text_password, hashed_password):
-    return bcrypt.checkpw(plain_text_password.encode('utf8'), hashed_password.encode('utf8'))
-
-@app.route('/signin.html', methods=['GET'])
-def signin_page():
-    return render_template('signin.html', last_updated=dir_last_updated('server/static'))
-
-@app.route('/signup.html', methods=['GET'])
-def signup_page():
-    return render_template('signup.html', last_updated=dir_last_updated('server/static'))
-
-@app.route('/profile.html', methods=['GET'])
-def profile_page():
-    return render_template('profile.html', last_updated=dir_last_updated('server/static'))
-
-@app.route('/api/1.0/signin', methods=['POST'])
-def signin():
-    form = request.form.to_dict()
-    email = form.get('email', None) 
-    password = form.get('password', None)
-
-    user = get_user(email)
-    if not user:
-        return jsonify({"error": "Bad username"}), 401
-
-    if not check_password(password, user["password"]):
-        return jsonify({"error": "Bad password"}), 401
-
-    access_token = create_access_token(identity=user["name"])
-    return {
-        "access_token": access_token,
-        "access_expired": 3600,
-        "user": {
-            "id": user["id"],
-            "rovider": 'native',
-            "name": user["name"],
-            "email": email,
-            "picture": ""
-        }
-    }
-
-@app.route('/api/1.0/signup', methods=['POST'])
-def signup():
-    form = request.form.to_dict()
-    name = form.get('name', None)
-    email = form.get('email', None) 
-    password = form.get('password', None)
-    encrypted_password = get_hashed_password(password)
-
-    user = get_user(email)
-    if user:
-        return jsonify({"error": "User already existed"}), 401
-
-    access_token = create_access_token(identity=name)
-    user_id = create_user('native', email, encrypted_password, name, access_token, 2592000)
-    return {
-        "access_token": access_token,
-        "access_expired": 3600,
-        "user": {
-            "id": user_id,
-            "rovider": 'native',
-            "name": name,
-            "email": email,
-            "picture": ""
-        }
-    }
-
-@app.route('/api/1.0/profile', methods=['GET'])
-@jwt_required()
-def protected():
-    current_user = get_jwt_identity()
-    print(current_user)
-    return f"Welcome! {current_user}"
-
-@app.route('/')
-@app.route('/dashboard')
-def dashboard():
-    return render_template('dashboard.html', last_updated=dir_last_updated('server/static'))
-
-from time import sleep
-@app.route('/sleep')
-def long_running():
-    sleep(10)
-    return "Wake up!"
-
-@app.route('/api/1.0/user/behavior/<date>')
-def api_get_user_behavior(date):
-    data = get_user_behavior_by_date(date)
-    if (data):
-        return {
-            "behavior_count": [data['view_count'], data['view_item_count'], data['add_to_cart_count'], data['checkout_count']],
-            "user_count": [data['unique_user_count'], data['new_user_count'], data['return_user_count']]
-        }
-    else:
-        return {
-            "behavior_count": [0, 0, 0, 0],
-            "user_count": [0, 0, 0]
-        }
-
-def is_integer(n):
-    try:
-        float(n)
-    except ValueError:
-        return False
-    else:
-        return float(n).is_integer()
+@app.route('/admin/recommendation.html')
+def product_recommendation_page():
+    res = get_products(100, 0, {"source": "amazon"},)
+    return render_template('product_recommendation.html', products = [{"id": p["id"], "title": p["title"]} for p in res["products"]])
 
 def find_product(category, paging):
     if (category == 'all') :
@@ -230,10 +118,6 @@ def allowed_file(filename):
     return '.' in filename and \
            filename.rsplit('.', 1)[1] in ALLOWED_EXTENSIONS
 
-@app.route('/admin/product.html', methods=['GET'])
-def admin_product():
-    return render_template('product_create.html')
-
 def save_file(folder, file):
     folder_root = app.root_path + app.config['UPLOAD_FOLDER']
     folder_path = folder_root + '/' + folder
@@ -293,8 +177,3 @@ def api_create_product():
 
     create_product(product, variants)
     return "Ok"
-
-@app.route('/recommendation')
-def product_recommendation():
-    res = get_products(100, 0, {"source": "amazon"},)
-    return render_template('product_recommendation.html', products = [{"id": p["id"], "title": p["title"]} for p in res["products"]])
