+++ App/server/utils/util.py
@@ -1,0 +1,5 @@
+import os
+def dir_last_updated(folder):
+    return str(max(os.path.getmtime(os.path.join(root_path, f))
+                   for root_path, dirs, files in os.walk(folder)
+                   for f in files))
